% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_reprogramming_blockers.R
\name{flag_reprogramming_blockers}
\alias{flag_reprogramming_blockers}
\title{Flag compounds that block beta cell reprogramming}
\usage{
flag_reprogramming_blockers(
  forward_results,
  reverse_results,
  toxic_drugs = c("streptozotocin", "alloxan", "STZ")
)
}
\arguments{
\item{forward_results}{Dataframe from query_cmap()}

\item{reverse_results}{Dataframe from query_cmap_reverse()}

\item{toxic_drugs}{Character vector of known beta cell toxic drugs}
}
\value{
A dataframe with a 'flag' column marking dangerous compounds
}
\description{
Flag compounds that block beta cell reprogramming
}
