% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_cmap.R
\name{query_cmap}
\alias{query_cmap}
\title{Query CMap/LINCS database with a gene signature}
\usage{
query_cmap(signature, db = "lincs", n_top = 100)
}
\arguments{
\item{signature}{A list with upgenes and downgenes (from build_signature)}

\item{db}{Database to query, either "lincs" or "lincs2" (default "lincs")}

\item{n_top}{Number of top results to return (default 100)}
}
\value{
A dataframe of top drug candidates with scores
}
\description{
Query CMap/LINCS database with a gene signature
}
