% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_cmap_reverse.R
\name{query_cmap_reverse}
\alias{query_cmap_reverse}
\title{Query CMap/LINCS database in reverse to identify harmful compounds}
\usage{
query_cmap_reverse(signature, db = "lincs", n_top = 100)
}
\arguments{
\item{signature}{A list with upgenes and downgenes (from build_signature)}

\item{db}{Database to query, either "lincs" or "lincs2" (default "lincs")}

\item{n_top}{Number of top results to return (default 100)}
}
\value{
A dataframe of compounds that would worsen beta cell state
}
\description{
Query CMap/LINCS database in reverse to identify harmful compounds
}
