% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReprogDrugR-package.R
\docType{package}
\name{ReprogDrugR-package}
\alias{ReprogDrugR}
\alias{ReprogDrugR-package}
\title{ReprogDrugR: Drug Reprogramming Screening for Beta Cell Regeneration}
\description{
Identifies drug candidates to reprogram senescent beta cells.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ZHANGJIDuhramUniversity/ReprogDrugR}
  \item Report bugs at \url{https://github.com/ZHANGJIDuhramUniversity/ReprogDrugR/issues}
}

}
\author{
\strong{Maintainer}: Ji Zhang \email{2742673684zj@gmail.com}

}
\keyword{internal}
